fix: Improve type parsing for parameter and return types
I may be wrong, but it seems like function parameter types and
return types are specified with the same syntax. I tried to simplify
by creating a scope I'm calling "qualified types" - which could
include the pointer operator, a lifetime, and/or if the item is
mutable. Seems to work for now, but I'm sure I'll run up against
an edge case soon.