feat: Add handler for format on save
This commit adds a call to `rustfmt` when a user saves. The method
returns a Promise to have the issue checker await the reformat (so
that highlighted ranges don't change). Unfortunately, Nova seems to
pick up on the altered document *after* it has generated issues,
and thus highlights are removed. Not the desired behavior.