feat: Add update script for Rust Analyzer binary
This script downloads the Rust Analyzer binary if it isn't installed
yet or if it doesn't match the current version, which will be passed
in as an argument. Using a shell script is much easier than spinning
up a bunch of processes for things that can only be done via command
line. The binary is unzipped alongside this script in the bin folder.