feat: Add function to run the update script and call from main
This commit adds a function that gets the latest version of Rust
Analyzer from GitHub, and passes this version when running the
update script. I can't run this whole process in development mode,
so I can only hope it runs as expected. Because I'm making a request
to GitHub, I needed to add the requests entitlement.