feat: Replace rust-analyzer binary after downloading update
This is another feature I can't quite test in development mode
(due to extension restarts on file changes), so I'm hoping it goes
well when released. After downloading an update for Rust Analyzer,
the extension then stops the language server process and overwrites
the binary before starting the server again. Using the Nova API to
do the file remove/move meant adding the entitlement for both
reading and writing files.