fix: Restart language server if it crashes
Instead of simply logging a server crash, this commit adds a user
notification that enables them to restart the server. Also, the
start method now checks that the Rust Analyzer binary exists before
trying to open a connection (this should only happen right after
a user installs the extension and tries to use it the first time).