refactor: Fix syntax naming
This commit adopts the Nova convention of starting a scope name with
the language name, instead of the Sublime convention of placine the
language name at the end of the scope name. A second change is
using the Rust naming of 'attributes' instead of the Nova theme
name of 'processing'.