refactor: Change enum descriptions, remove some workspace config items
I discovered that Nova doesn't reconcile global and workspace
preferences, so I'm choosing to be pickier about what is configurable
in a workspace since I need to do the reconciling. I chose to leave
format-on-save as a global preference, but the issue checker will be
a workspace preference. I also switched to using nested arrays for
the enum values, so I could move descriptions to the
items themselves.