feat: Add utility function to reconcile global and local preferences
I thought Nova did this for me, but it's up to me. This could be
because some preferences might be better as additive between global
and local, instead of the local preference always overwriting the
global preference. Either way, I hope this function to save me
some boilerplate when I repeat a preference for a workspace.