refactor: Configure Rust Analyzer to generate issues
Up until now, I've used a class (RustIssueProvider) to run
`cargo check` and generate Nova Issues from the output because the
Nova LanguageClient wasn't reliably creating issues from Rust
Analyzer. With Nova 9, this and other LanguageClient issues seem
resolved, so I'ved decided to rely on Rust Analyzer for issues.
This means recreating the client for any preference change on the
lint command and requiring Nova 9 for the next extension version.
I'm leaving the RustIssueProvider class for now, but have removed
its usage when documents are saved.