feat: Revise task template to use resolved tasks/actions
Taking a slightly different tack and am going to have the same
template able to be modified if a user wants it for running tests
or benchmarks. I also discovered that resolving arguments is
necessary so that Nova doesn't try to provide an empty ('') argument
to the task process. Finally, resolving a task will allow me to
insert environment variables that may be needed. The tasks are
currently not resolving in Nova 9.6 or 10b4, but I'm hoping a fix
will come soon.