feat: Add environment variables to Rust Analyzer
This commit pulls changes to environment variables preferences into
the LanguageClient and adds them to the initialization of Rust
Analyzer. As part of this commit, I needed to add a few extra
checks to the function that converts the string array to an object.
Apparently nulls are sent? There was some sort of null error in
the console that went away after I added a value.