feat: Add function to split an argument string for processes
To keep it easy for users, arguments to a CLI command are gathered
with a simple string input field. Processes, however, expect a nice,
tidy array of argument strings. This commit adds a function that
splits on space characters unless it's within quotes. The Cargo
task template and Rust Analyzer lint command use this function for
preparing preference inputs as process arguments.