fix: Use correct path for `rustfmt.toml`
When I added finding the `Cargo.toml` to get the Rust edition
for formatting, I carelessly didn't return the path to the
`rustfmt.toml` if found but instead its parent directory. This
commit returns the path to the file. It also fixes the use of
the deprecated `substr` function.

Issues: #43